backend_v2/requirements.txt
apify-client>=1.7.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.8.0
supabase>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
